Extract survey and feedback api endpoints
