+++ controllers/api/index.js
@@ -2,8 +2,6 @@
 const express = require('express');
 const Sentry = require('@sentry/node');
 
-const { sanitise } = require('../../common/sanitise');
-const { Feedback, SurveyAnswer } = require('../../db/models');
 const appData = require('../../common/appData');
 const { POSTCODES_API_KEY } = require('../../common/secrets');
 const { csrfProtection } = require('../../common/cached');
@@ -12,8 +10,6 @@ const { Client } = require('@ideal-postcodes/core-node');
 const postcodesClient = new Client({
     api_key: POSTCODES_API_KEY
 });
-
-const { validateFeedback, validateSurvey } = require('./schemas');
 
 const router = express.Router();
 
@@ -66,70 +62,11 @@ router.post('/address-lookup', csrfProtection, async (req, res) => {
 /**
  * API: Feedback endpoint
  */
-router.post('/feedback', async (req, res) => {
-    const validationResult = validateFeedback(req.body);
-
-    const messageSuccess = req.i18n.__('global.feedback.success');
-    const messageError = req.i18n.__('global.feedback.error');
-
-    if (validationResult.error) {
-        res.status(400).json({
-            status: 'error',
-            message: messageError,
-            err: validationResult.error.message
-        });
-    } else {
-        try {
-            const [result] = await Feedback.storeFeedback({
-                description: validationResult.value.description,
-                message: sanitise(validationResult.value.message)
-            });
-
-            res.json({
-                status: 'success',
-                message: messageSuccess,
-                result: result
-            });
-        } catch (storeError) {
-            res.status(400).json({
-                status: 'error',
-                message: messageError,
-                err: storeError.message
-            });
-        }
-    }
-});
+router.use('/feedback', require('./feedback'));
 
 /**
  * API: Survey endpoint
  */
-router.post('/survey', async (req, res) => {
-    const validationResult = validateSurvey(req.body);
-
-    if (validationResult.error) {
-        res.status(400).json({
-            status: 'error',
-            err: validationResult.error.message
-        });
-    } else {
-        try {
-            const [result] = await SurveyAnswer.createResponse({
-                choice: validationResult.value.choice,
-                path: validationResult.value.path,
-                message: sanitise(validationResult.value.message)
-            });
-
-            res.json({
-                status: 'success',
-                result: result
-            });
-        } catch (storeError) {
-            res.status(400).json({
-                status: 'error',
-                err: storeError.message
-            });
-        }
-    }
-});
+router.use('/survey', require('./survey'));
 
 module.exports = router;
