+++ controllers/api/lib/validate-feedback.js
@@ -1,0 +1,14 @@
+'use strict';
+const Joi = require('@hapi/joi16');
+
+module.exports = function validateFeedback(data = {}) {
+    const schema = Joi.object({
+        description: Joi.string().required(),
+        message: Joi.string().required()
+    });
+
+    return schema.validate(data, {
+        abortEarly: false,
+        stripUnknown: true
+    });
+};
