+++ controllers/api/lib/validate-feedback.test.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 'use strict';
-const { validateFeedback, validateSurvey } = require('./schemas');
+const validateFeedback = require('./validate-feedback');
 const faker = require('faker');
 
 function mapMessages(validationResult) {
@@ -20,24 +20,5 @@ test('validate feedback schema', () => {
     expect(mapMessages(validateFeedback({}))).toEqual([
         '"description" is required',
         '"message" is required'
-    ]);
-});
-
-test('validate survey schema', () => {
-    const valid = validateSurvey({
-        choice: 'yes',
-        path: '/some/path',
-        message: faker.lorem.words(25)
-    });
-
-    expect(valid.error).toBeUndefined();
-
-    const invalid = validateSurvey({
-        choice: 'not-a-valid-choice'
-    });
-
-    expect(mapMessages(invalid)).toEqual([
-        '"choice" must be one of [yes, no]',
-        '"path" is required'
     ]);
 });
